Mqtt5 fixes

+++ wire5/src/commonMain/kotlin/mqtt/wire5/control/packet/format/variable/property/ReceiveMaximum.kt
@@ -5,6 +5,7 @@ package mqtt.wire5.control.packet.format.variable.property
 import kotlinx.io.core.BytePacketBuilder
 import mqtt.wire.data.Type
 
-data class ReceiveMaximum(val maxQos1Or2ConcurrentMessages: UShort) : Property(0x21, Type.TWO_BYTE_INTEGER) {
-    override fun write(bytePacketBuilder: BytePacketBuilder) = write(bytePacketBuilder, maxQos1Or2ConcurrentMessages)
+data class ReceiveMaximum(val maxQos1Or2ConcurrentMessages: Int) : Property(0x21, Type.TWO_BYTE_INTEGER) {
+    override fun write(bytePacketBuilder: BytePacketBuilder) =
+        write(bytePacketBuilder, maxQos1Or2ConcurrentMessages.toUShort())
 }
