+++ wire5/src/commonMain/kotlin/mqtt/wire5/control/packet/format/variable/property/SessionExpiryInterval.kt
@@ -5,6 +5,6 @@ package mqtt.wire5.control.packet.format.variable.property
 import kotlinx.io.core.BytePacketBuilder
 import mqtt.wire.data.Type
 
-data class SessionExpiryInterval(val seconds: UInt) : Property(0x11, Type.FOUR_BYTE_INTEGER) {
-    override fun write(bytePacketBuilder: BytePacketBuilder) = write(bytePacketBuilder, seconds)
+data class SessionExpiryInterval(val seconds: Long) : Property(0x11, Type.FOUR_BYTE_INTEGER) {
+    override fun write(bytePacketBuilder: BytePacketBuilder) = write(bytePacketBuilder, seconds.toUInt())
 }
