+++ wire5/src/commonMain/kotlin/mqtt/wire5/control/packet/format/variable/property/SubscriptionIdentifier.kt
@@ -6,6 +6,7 @@ import kotlinx.io.core.BytePacketBuilder
 import mqtt.wire.data.Type
 import mqtt.wire.data.VariableByteInteger
 
-data class SubscriptionIdentifier(val value: UInt) : Property(0x0B, Type.VARIABLE_BYTE_INTEGER) {
-    override fun write(bytePacketBuilder: BytePacketBuilder) = write(bytePacketBuilder, VariableByteInteger(value).encodedValue())
+data class SubscriptionIdentifier(val value: Long) : Property(0x0B, Type.VARIABLE_BYTE_INTEGER) {
+    override fun write(bytePacketBuilder: BytePacketBuilder) =
+        write(bytePacketBuilder, VariableByteInteger(value.toUInt()).encodedValue())
 }
