+++ wire5/src/commonMain/kotlin/mqtt/wire5/control/packet/format/variable/property/TopicAliasMaximum.kt
@@ -5,6 +5,7 @@ package mqtt.wire5.control.packet.format.variable.property
 import kotlinx.io.core.BytePacketBuilder
 import mqtt.wire.data.Type
 
-data class TopicAliasMaximum(val highestValueSupported: UShort) : Property(0x23, Type.TWO_BYTE_INTEGER) {
-    override fun write(bytePacketBuilder: BytePacketBuilder) = write(bytePacketBuilder, highestValueSupported)
+data class TopicAliasMaximum(val highestValueSupported: Int) : Property(0x23, Type.TWO_BYTE_INTEGER) {
+    override fun write(bytePacketBuilder: BytePacketBuilder) =
+        write(bytePacketBuilder, highestValueSupported.toUShort())
 }
