+++ wire5/src/commonMain/kotlin/mqtt/wire5/control/packet/format/variable/property/WillDelayInterval.kt
@@ -5,6 +5,6 @@ package mqtt.wire5.control.packet.format.variable.property
 import kotlinx.io.core.BytePacketBuilder
 import mqtt.wire.data.Type
 
-data class WillDelayInterval(val seconds: UInt) : Property(0x18, Type.FOUR_BYTE_INTEGER, willProperties = true) {
-    override fun write(bytePacketBuilder: BytePacketBuilder) = write(bytePacketBuilder, seconds)
+data class WillDelayInterval(val seconds: Long) : Property(0x18, Type.FOUR_BYTE_INTEGER, willProperties = true) {
+    override fun write(bytePacketBuilder: BytePacketBuilder) = write(bytePacketBuilder, seconds.toUInt())
 }
